Change position coordinates and time to doubles.

+++ src/main/java/timelessodyssey/Game.java
@@ -46,7 +46,7 @@ public class Game {
 
     private void start() throws IOException, InterruptedException, URISyntaxException, FontFormatException {
         int FPS = 30;
-        int frameTime = 1000 / FPS;
+        double frameTime = 1000.0 / FPS;
 
         while (this.state != null) {
             long startTime = System.currentTimeMillis();
@@ -54,7 +54,7 @@ public class Game {
             state.step(this, gui, startTime);
 
             long elapsedTime = System.currentTimeMillis() - startTime;
-            long sleepTime = frameTime - elapsedTime;
+            long sleepTime = (long)frameTime - elapsedTime;
 
             if (sleepTime > 0) Thread.sleep(sleepTime);
         }
